password update route name changement reuse the password recovery component to use it for the first connection password update. Rename the forgotPassword folders to Password. Added condition for workspaces and messages.
