+++ src/Components/Context/UserContext.tsx
@@ -2,6 +2,7 @@
 import { createContext, useEffect, useState } from 'react'
 import jwt_decode from 'jwt-decode'
 import { withRouter } from 'react-router-dom'
+import { useToasts } from 'react-toast-notifications'
 import { iUsers, iTokenDecrypted } from '../../Interfaces/UsersInterfaces'
 
 export const UserContext: any = createContext(null)
@@ -14,6 +15,7 @@ const UserProvider = withRouter((props) => {
   const { history, children } = props
   const [token, setToken] = useState<string | null>(null)
   const [userInfos, setUserInfos] = useState<iUsers | null>(null)
+  const { addToast } = useToasts()
 
   useEffect(() => {
     if (tokenInLocalStorage) {
@@ -28,11 +30,22 @@ const UserProvider = withRouter((props) => {
   }
 
   const addUser = (userToken: string) => {
-    history.push('/')
     localStorage.setItem('token', userToken)
     setToken(userToken)
     const userData: iTokenDecrypted = jwt_decode(userToken)
+    if (userData?.first_connection === true) {
+      addToast(
+        'Merci de cliquer sur le lien de changement de mot de passe qui vous a été envoyer par mail',
+        {
+          appearance: 'error',
+          autoDismiss: false,
+        },
+      )
+      return history.push('/login')
+    }
     // check if the token isn't expired
+    history.push('/')
+
     if (userData && userData.exp) {
       const now = Date.now()
       if (now > userData.exp * 1000) {
