+++ src/Pages/Password/MailSentStyle.tsx
