+++ src/Pages/Password/NewPasswordStyle.tsx
